pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bloginator"
version = "1.0.0"
description = "Content generation from your own writing corpus"
readme = "README.md"
requires-python = ">=3.10,<3.15"
license = {text = "MIT"}
authors = [
    {name = "Matt Bordenet", email = "matt@bordenet.com"},
]
keywords = [
    "content-generation",
    "rag",
    "llm",
    "voice-preservation",
    "local-llm",
    "chromadb",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
]

dependencies = [
    "click>=8.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "striprtf>=0.0.26",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "ollama>=0.1.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.2",
    "cryptography>=43.0.1",  # Security: CVE fixes
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "pydocstyle>=6.3.0",
    "bandit>=1.7.5",
    "types-requests>=2.31.0",
]

local = [
    "ollama>=0.1.0",
]

cloud = [
    "anthropic>=0.7.0",
    "openai>=1.3.0",
]

web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
]

all = [
    # Dev dependencies
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
    "pydocstyle>=6.3.0",
    "bandit>=1.7.5",
    "types-requests>=2.31.0",
    # Local LLM
    "ollama>=0.1.0",
    # Cloud LLMs
    "anthropic>=0.7.0",
    "openai>=1.3.0",
    # Web UI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
]

[project.scripts]
bloginator = "bloginator.cli.main:cli"

[project.urls]
Homepage = "https://github.com/bordenet/bloginator"
Documentation = "https://github.com/bordenet/bloginator/blob/main/README.md"
Repository = "https://github.com/bordenet/bloginator"
Issues = "https://github.com/bordenet/bloginator/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["bloginator*"]
exclude = ["tests*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PTH",  # flake8-use-pathlib
]
ignore = [
    "E501",  # line too long (handled by black)
    "B904",  # Within except clause, raise with from err - intentional in error handlers
    "B008",  # Do not perform function calls in argument defaults - FastAPI pattern
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "S101",   # Allow assert in tests
    "ARG",    # Allow unused arguments in tests
    "FBT",    # Allow boolean trap in tests
]

[tool.ruff.lint.isort]
known-first-party = ["bloginator"]
force-single-line = false
lines-after-imports = 2

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2
src_paths = ["src", "tests"]
known_first_party = ["bloginator"]
skip_gitignore = true

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.ui.*"
disallow_untyped_defs = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "bloginator.cli.serve"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.cli.error_reporting"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.cli.search"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.generation.llm_mock"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.utils.parallel"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "bloginator.cli.extract_utils"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_ollama: marks tests that require Ollama to be running",
    "requires_api_key: marks tests that require API keys",
    "e2e: marks tests as end-to-end tests",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as performance benchmarks",
    "performance: marks tests as performance benchmarks",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# Pydocstyle configuration
[tool.pydocstyle]
convention = "google"
add_ignore = ["D100", "D104"]  # Don't require docstrings for packages and __init__.py
match_dir = "^(?!tests|venv|build|dist).*"
